+++ src/implementation/landmarks.cpp
@@ -12,8 +12,8 @@ using namespace std;
 
 struct comparator{
     constexpr bool operator()(
-            pair<int, double> const& a,
-            pair<int, double> const& b)
+            pair<int, float> const& a,
+            pair<int, float> const& b)
     const noexcept
     {
         return a.second > b.second;
@@ -36,13 +36,13 @@ vector<landmarksStruct> landmarks::initLandmarks(vector<long long> nodeIDs, adjL
 
 adjListCollection landmarks::reverseAdjListCollection(adjListCollection &adjCol){
     adjListCollection reversedAdjCol;
-    reversedAdjCol.adjlst.resize(adjCol.idSoFar, vector<pair<int, double>>{});
+    reversedAdjCol.adjlst.resize(adjCol.idSoFar, vector<pair<int, float>>{});
     for(int i = 0; i < adjCol.adjlst.size(); i++){
-        vector<pair<int, double>> currentEdges = adjCol.adjlst[i];
+        vector<pair<int, float>> currentEdges = adjCol.adjlst[i];
         for(int j = 0; j < currentEdges.size(); j++) {
-            pair<int, double> temp = currentEdges[j];
+            pair<int, float> temp = currentEdges[j];
             int newSource = temp.first;
-            double newDist = temp.second;
+            float newDist = temp.second;
             reversedAdjCol.adjlst[newSource].emplace_back(make_pair(i, newDist));
         }
     }
@@ -50,16 +50,15 @@ adjListCollection landmarks::reverseAdjListCollection(adjListCollection &adjCol)
     return reversedAdjCol;
 }
 
-vector<landmarksStruct>
-landmarks::initLandmarks(int amount, adjListCollection &adjListCollection, const string& landmarkSelection) {
+vector<landmarksStruct> landmarks::initLandmarks(int amount, adjListCollection &adjListCollection, const string& landmarkSelection) {
     struct adjListCollection reversedAdjListCollection = reverseAdjListCollection(adjListCollection);
     vector<landmarksStruct> resultVector;
     int highestNbr = adjListCollection.idSoFar;
 
     int randomNode = rand() % highestNbr;
-    const double INF = std::numeric_limits<double>::infinity();
+    const float INF = std::numeric_limits<float>::infinity();
     //cout << "https://www.openstreetmap.org/node/" << adjListCollection.intIdToLongID[randomNode] << endl;
-    vector<vector<double>> markDistanceVectors;
+    vector<vector<float>> markDistanceVectors;
     for (int i = 0; i < amount+1; ++i) {
 
         if(i!=0){cout << "https://www.openstreetmap.org/node/" << adjListCollection.intIdToLongID[randomNode] << endl;}
@@ -73,16 +72,16 @@ landmarks::initLandmarks(int amount, adjListCollection &adjListCollection, const
 
         //cout << "https://www.openstreetmap.org/node/" << adjListCollection.intIdToLongID[randomNode] << "#map=8/56.216/12.816"<< endl;
         if(landmarkSelection == "euclidDistance"){
-            vector<double> euclidDistVector;
+            vector<float> euclidDistVector;
             euclidDistVector.resize(highestNbr,0);
-            double sourceX = adjListCollection.xCoord[randomNode];
-            double sourceY = adjListCollection.yCoord[randomNode];
+            float sourceX = adjListCollection.xCoord[randomNode];
+            float sourceY = adjListCollection.yCoord[randomNode];
             for (int j = 0; j < highestNbr; ++j) {
                 //if(j % 1000 == 0)
                 //    cout << "euclidloop: " << j << endl;
-                double targetX = adjListCollection.xCoord[j];
-                double targetY = adjListCollection.yCoord[j];
-                double euclidDistToJ = adjacencyList::euclidDistance(sourceX, sourceY, targetX, targetY);
+                float targetX = adjListCollection.xCoord[j];
+                float targetY = adjListCollection.yCoord[j];
+                float euclidDistToJ = adjacencyList::euclidDistance(sourceX, sourceY, targetX, targetY);
                 euclidDistVector[j]=euclidDistToJ;
             }
             markDistanceVectors.emplace_back(euclidDistVector);
@@ -90,15 +89,13 @@ landmarks::initLandmarks(int amount, adjListCollection &adjListCollection, const
         resultVector.emplace_back(landmarksStruct);
 
 
-        double longestDistToClosestMark = 0.0;
+        float longestDistToClosestMark = 0.0;
         for (int j = 0; j <= highestNbr; ++j) { //loops over all nodes
-            //if(j % 1000 == 0)
-            //    cout << "longestDistToClosestMark: " << j << endl;
-            double closestMark = INF;
+            float closestMark = INF;
             //for (vector<double> distVec : markDistanceVectors){
             for (int k = 0; k < resultVector.size(); ++k) {
-                double distToCandidate =0;
-                double distToCandidateDijkstra;
+                float distToCandidate =0;
+                float distToCandidateDijkstra;
                 if(landmarkSelection == "euclidDistance"){
                     distToCandidate = markDistanceVectors[k][j];
                     distToCandidateDijkstra = resultVector[k].distanceVec[j];
@@ -109,8 +106,6 @@ landmarks::initLandmarks(int amount, adjListCollection &adjListCollection, const
                 else {
                     cout << "Unknown landmarkSelection provided: " << landmarkSelection << endl;
                 }
-                //if(j % 1000 == 0)
-                //    cout << "looping over landmarks: " << j << endl;
                 //double distToCandidate = distVec[j];
                 if (distToCandidate != INF && distToCandidateDijkstra != INF && distToCandidate < closestMark)
                     closestMark = distToCandidate;
@@ -119,23 +114,10 @@ landmarks::initLandmarks(int amount, adjListCollection &adjListCollection, const
                 longestDistToClosestMark = closestMark;
                 randomNode = j;
             }
-            //if(j % 1000 == 0)
-            //    cout << "last thing of the loop: " << j << endl;
         }
         //if(i==0) markDistanceVectors.pop_back();
     }
     resultVector.erase(resultVector.begin());
-    //cout<<"size of result vector: " << resultVector.size() << endl;
-    //for (landmarksStruct i:resultVector) {
-    //cout << "lmk: " << i.nodeID << endl;
-    //}
-
-    //recast to long longs for calling initLandmarks
-    //vector<long long> longIdsOfLandmarks;
-    //for (int id:landmarkIDS) {
-    //    long long longID = adjacencyList::getLongID(adjListCollection,id);
-    //    longIdsOfLandmarks.push_back(longID);
-    //}
     return resultVector;
 }
 
@@ -143,22 +125,22 @@ landmarks::initLandmarks(int amount, adjListCollection &adjListCollection, const
 spResultStruct landmarks::ALTShortestPath(int source, int dest, adjListCollection &adjCol) {
     //Choose which landmark is best for this problem
     landmarksStruct landmark = choseLandmarks(source, dest, adjCol);
-    const double INF = std::numeric_limits<double>::infinity();
+    const float INF = std::numeric_limits<float>::infinity();
     int sizeOfGraph = adjCol.idSoFar; //Amount of nodes in graph
     //initialize distance from source to everything to infinity, and source to 0
-    vector<double> distance(sizeOfGraph, INF);
+    vector<float> distance(sizeOfGraph, INF);
     distance[source] = 0;
     //has the node been seen vector
     vector<bool> nodeSeen(sizeOfGraph, false);
     //The pi array containing the path from source to destination
     vector<int> prevNode(sizeOfGraph, -1);
     //heap of nodes to evaluate
-    priority_queue<pair<int, double>, vector<pair<int, double>>, comparator> minHeap;
+    priority_queue<pair<int, float>, vector<pair<int, float>>, comparator> minHeap;
 
     minHeap.push(make_pair(source, 0.0));
     while (!minHeap.empty()){
         //pop the top element
-        pair<int,double> head = minHeap.top();
+        pair<int,float> head = minHeap.top();
         minHeap.pop();
         int headId = head.first;
 
@@ -170,8 +152,8 @@ spResultStruct landmarks::ALTShortestPath(int source, int dest, adjListCollectio
         auto connectedNodes = adjCol.adjlst[headId];
         for(auto const &i: connectedNodes){ //Check adjacent nodes
             int node = i.first;
-            double weight = i.second;
-            double heuristIntermediate = calcHeuristicDistance(node, dest, landmark);
+            float weight = i.second;
+            float heuristIntermediate = calcHeuristicDistance(node, dest, landmark);
             //relaxation step, in ALT we add the heuristic weight in to consideration
             if(!nodeSeen[node] && (distance[headId] + weight) < distance[node]){
                 distance[node] = distance[headId]+weight;
@@ -188,10 +170,10 @@ spResultStruct landmarks::ALTShortestPath(int source, int dest, adjListCollectio
 
 landmarksStruct landmarks::choseLandmarks(int source, int dest, adjListCollection &collection) {
     int bestBounding;
-    double bestBound = 0;
+    float bestBound = 0;
     //Iterate over all landmarks and choose the one with the best bound
     for(int i = 0; i < collection.landmarksStructs.size(); i++) {
-        double lowerBound = calcHeuristicDistance(source, dest,collection.landmarksStructs[i]);
+        float lowerBound = calcHeuristicDistance(source, dest,collection.landmarksStructs[i]);
         if(bestBound == 0 || lowerBound > bestBound) {
             bestBounding = i;
             bestBound = lowerBound;
@@ -200,15 +182,15 @@ landmarksStruct landmarks::choseLandmarks(int source, int dest, adjListCollectio
     return collection.landmarksStructs[bestBounding];
 }
 
-double landmarks::calcHeuristicDistance(int source, int target, landmarksStruct &currLandmark) {
+float landmarks::calcHeuristicDistance(int source, int target, landmarksStruct &currLandmark) {
     //From source/Target to landmark
-    double distFromSourceToLandmark = currLandmark.reversedDistanceVec[source];
-    double distFromTargetToLandmark = currLandmark.reversedDistanceVec[target];
-    double toLandmark = distFromSourceToLandmark - distFromTargetToLandmark;
+    float distFromSourceToLandmark = currLandmark.reversedDistanceVec[source];
+    float distFromTargetToLandmark = currLandmark.reversedDistanceVec[target];
+    float toLandmark = distFromSourceToLandmark - distFromTargetToLandmark;
     //From landmark to source/Target
-    double distFromLandmarkToSource = currLandmark.distanceVec[source];
-    double distFromLandmarkToTarget = currLandmark.distanceVec[target];
-    double fromLandmark = distFromLandmarkToTarget - distFromLandmarkToSource;
+    float distFromLandmarkToSource = currLandmark.distanceVec[source];
+    float distFromLandmarkToTarget = currLandmark.distanceVec[target];
+    float fromLandmark = distFromLandmarkToTarget - distFromLandmarkToSource;
 
     //handle the case where either source or target is not in the same graph as the landmark
     bool isLowerBFromInfinite =  !isfinite(fromLandmark);
@@ -222,7 +204,7 @@ double landmarks::calcHeuristicDistance(int source, int target, landmarksStruct 
     if (isLowerBToInfinite){
         return fromLandmark;
     }
-    double largestLowerbound = max(toLandmark, fromLandmark);
+    float largestLowerbound = max(toLandmark, fromLandmark);
     return largestLowerbound;
 
 }
